Refactor Snake and Apple classes for improved position handling
- Updated position attributes in Snake and Apple classes from body_pos to body for consistency.
- Ensured position calculations in draw methods use int casting for accurate rendering.
- Refactored move and eat methods in Snake class to align with new body attribute structure.

+++ src/object.py
@@ -37,8 +37,8 @@ class GreenApple(Apple):
 
     def draw(self, screen):
         apple_rect = pygame.Rect(
-            self.pos.x * CELL_SIZE,
-            self.pos.y * CELL_SIZE,
+            int(self.pos.x * CELL_SIZE),
+            int(self.pos.y * CELL_SIZE),
             CELL_SIZE,
             CELL_SIZE
         )
@@ -58,8 +58,8 @@ class RedApple(Apple):
 
     def draw(self, screen):
         apple_rect = pygame.Rect(
-            self.pos.x * CELL_SIZE,
-            self.pos.y * CELL_SIZE,
+            int(self.pos.x * CELL_SIZE),
+            int(self.pos.y * CELL_SIZE),
             CELL_SIZE,
             CELL_SIZE
         )
